+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @electron-boot/framework
 
+## 2.1.4
+
+### Patch Changes
+
+- 增加注册channel日志
+
 ## 2.1.3
 
 ### Patch Changes