+++ packages/pouchdb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @electron-boot/pouchdb
 
+## 2.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @electron-boot/framework@2.1.4
+
 ## 2.1.2
 
 ### Patch Changes